+++ src/main/kotlin/com/example/kbocombo/domain/player/vo/HandType.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.domain
+package com.example.kbocombo.domain.player.vo
 
 enum class HandType {
     LEFT,
