+++ src/main/kotlin/com/example/kbocombo/domain/player/vo/WebId.kt
@@ -1,4 +1,4 @@
-package com.example.kbocombo.domain
+package com.example.kbocombo.domain.player.vo
 
 @JvmInline
 value class WebId(
